refactor: User -> Catalog 수정
