+++ CatalogService/.gitattributes
