+++ CatalogService/build.gradle
