+++ CatalogService/src/main/resources/application.properties
@@ -1,6 +1,6 @@
-spring.application.name=UserService
+spring.application.name=CatalogService
 
-server.port = 8081
+server.port = 8086
 
 # DB connect
 spring.datasource.url=${SPRING_DATASOURCE_URL}
